Home bisa ambil data user yang login dan ditampilin
Nambah junction table JadwalTayang dan rubah relasi dan migrasi

+++ tubez/lib/entity/User.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 class User{
-  BigInt? id;
+  int? id;
   String username;
   String password;
   String tanggalLahir;
